PostreSQL Connection Created!

+++ Core/SednaReservationAPI.Domain/Entities/Review.cs
@@ -10,7 +10,7 @@ namespace SednaReservationAPI.Domain.Entities
     {
         public Guid HotelId { get; set; }
         public Guid UserId { get; set; }
-        public int Rating { get; set; }
+        public float Rating { get; set; }
         public string? Comment { get; set; }
     }
 }
